// Portal.Portal.Web/src/app/shared/infrastructure/PortalHttpClient/model/timeOffRequestType.ts
/**
 * Demo API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type TimeOffRequestType = 'Vacation' | 'DayOff' | 'SickLeave' | 'ParentalLeave' | 'BusinessTrip' | 'UnpaidVacation' | 'RemoteDay' | 'Custom';

export const TimeOffRequestType = {
    Vacation: 'Vacation' as TimeOffRequestType,
    DayOff: 'DayOff' as TimeOffRequestType,
    SickLeave: 'SickLeave' as TimeOffRequestType,
    ParentalLeave: 'ParentalLeave' as TimeOffRequestType,
    BusinessTrip: 'BusinessTrip' as TimeOffRequestType,
    UnpaidVacation: 'UnpaidVacation' as TimeOffRequestType,
    RemoteDay: 'RemoteDay' as TimeOffRequestType,
    Custom: 'Custom' as TimeOffRequestType
};